pachyderm/pachyderm/pachy_source/requirements.txt
opencv-python
cmake
dlib
tensorflow == 2.3.1
scipy == 1.4.1 
numpy == 1.18.5
Pandas == 1.1.5
tqdm == 4.62.0
h5py == 2.10.0
hydra-core
scikit-learn == 0.24.2
albumentations
omegaconf
ubiops
